+++ tests/test_manhole.py
@@ -92,6 +92,17 @@ def test_install_once():
             wait_for_strings(proc.read, TIMEOUT, 'ALREADY_INSTALLED')
 
 
+def test_install_twice_not_strict():
+    with TestProcess(sys.executable, HELPER, 'test_install_twice_not_strict') as proc:
+        with dump_on_error(proc.read):
+            wait_for_strings(proc.read, TIMEOUT, 'Not patching os.fork and os.forkpty. Oneshot activation is done by signal')
+            proc.signal(signal.SIGUSR2)
+            wait_for_strings(proc.read, TIMEOUT, '/tmp/manhole-')
+            uds_path = re.findall(r"(/tmp/manhole-\d+)", proc.read())[0]
+            wait_for_strings(proc.read, TIMEOUT, 'Waiting for new connection')
+            assert_manhole_running(proc, uds_path, oneshot=True)
+
+
 @mark.skipif(is_module_available('eventlet'), reason="evenlet can't deal with extra threads at process exit")
 def test_daemon_connection():
     with TestProcess(sys.executable, HELPER, 'test_daemon_connection') as proc: