+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 ------------------
 
 * Added support for installing the manhole via the ``PYTHONMANHOLE`` environment variable.
+* Added a ``strict`` install option. Set it to false to avoid getting the ``AlreadyInstalled`` exception.
 
 1.0.0 (2014-10-13)
 ------------------